+++ src/app/manage/[object]/add-categories-creator.tsx
@@ -14,6 +14,7 @@ import { Loader } from '@/components/common/loader';
 
 export function AddCategoriesCreator() {
     const [isLoadingObject, setIsLoadingObject] = useState<boolean>(false);
+    const [hasImage, setHasImage] = useState<boolean>(false);
     const queryParams = useSearchParams();
     const objectParam = queryParams.get('object');
     const editParam = queryParams.get('edit');
@@ -23,7 +24,7 @@ export function AddCategoriesCreator() {
     const formRef = useRef(null);
     const { setIsLoading, reFetch, needsToRefreshForm } = useContext(ManageCreatorContext);
 
-    const { categoryName, onCategoryNameChange, image, onImageChange, onSubmit, loadForm, resetForm } = useCategoriesCreator(setIsLoading, reFetch, formRef, editParam, router);
+    const { categoryName, onCategoryNameChange, image, onImageChange, onSubmit, loadForm, resetForm, onDeleteImage } = useCategoriesCreator(setIsLoading, reFetch, formRef, editParam, router);
 
     useEffect(() => {
         if (objectParam === 'categories' && editParam) {
@@ -34,6 +35,7 @@ export function AddCategoriesCreator() {
                 .getCategoryImageEnd(editParam)
                 .then(img => {
                     file = new File([img.blob], `${editParam}-category-image`, { type: img.blob.type });
+                    setHasImage(true);
                 })
                 .catch(err => {
                     if (err.message !== 'Image does not exist!') toastError(`Error: ${err.message}`);
@@ -61,6 +63,7 @@ export function AddCategoriesCreator() {
             <InputString label={'Category name'} name={'category-name'} value={categoryName} onChange={onCategoryNameChange} disabled={disabled} />
             <InputImage label={'Category image'} name={'category-image'} image={image} onChange={onImageChange} />
             <div className={styles['navigation-box']}>
+                {image && editParam && hasImage ? <SubmitButton label={'Delete image'} onSubmit={onDeleteImage} size="small" /> : <div></div>}
                 <SubmitButton label={'Submit'} size="small" onSubmit={onSubmit} />
             </div>
         </form>